requirements.txt
Flask
cohere
fpdf
pandas
openpyxl
python-dotenv